app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
streamlit
langgraph
langchain
langchain_groq
langchain-google-genai
grandalf
python_dotenv
langchain_community
fastapi
sseclient